feat: Add Order and OrderRegel CRUD functionality

+++ app/DataGrids/Settings/OrderDataGrid.php
@@ -54,7 +54,7 @@ class OrderDataGrid extends DataGrid
 
         $this->addColumn([
             'index'      => 'total_price',
-            'type'       => 'decimal',
+            'type'       => 'float',
             'label'      => 'Totale prijs',
             'searchable' => true,
             'filterable' => true,
